nlp/requirements.txt
https://github.com/eCAPTION/ecaption_utils/archive/refs/heads/main.zip
PyFunctional
spacy
python-dotenv
